# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/email/exist": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "是否已经注册过该 email",
                "parameters": [
                    {
                        "maxLength": 50,
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/reply.ParamExistEmail"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/email/send": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "发送验证码(邮件)",
                "parameters": [
                    {
                        "description": "email",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamSendEmail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2006:邮箱已经注册 2003:邮件发送过于频繁，请稍后再试",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/file/avatar/account": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "file"
                ],
                "summary": "更新账户头像",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "description": "文件及账户信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamUploadAccountAvatar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足 8001:存储失败",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamUploadAvatar"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/file/delete": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "删除文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "文件ID",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamDeleteFile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 8002:文件不存在 8003文件删除失败",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/file/details": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "获取文件详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "文件及账户信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamGetFileDetailsByID"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamFile"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/file/getFiles": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "获取关系文件列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "关系ID",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamGetRelationFile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足 8001:存储失败",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamGetRelationFile"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/file/publish": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "file"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "account_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "relation_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 8001:存储失败",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamPublishFile"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.State": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码，0：成功；否则失败",
                    "type": "integer"
                },
                "data": {
                    "description": "数据，失败时返回空"
                },
                "msg": {
                    "description": "状态的具体描述",
                    "type": "string"
                }
            }
        },
...
        "reply.ParamExistEmail": {
            "type": "object",
            "properties": {
                "exist": {
                    "description": "是否存在",
                    "type": "boolean"
                }
            }
        },
        "reply.ParamFile": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "create_at": {
                    "type": "string"
                },
                "file_id": {
                    "type": "integer"
                },
                "file_name": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "file_type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "reply.ParamGetRelationFile": {
            "type": "object",
            "properties": {
                "file_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reply.ParamFile"
                    }
                }
            }
        },
...
        "reply.ParamPublishFile": {
            "type": "object",
            "properties": {
                "create_at": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "file_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "reply.ParamUploadAvatar": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
...
        "request.ParamDeleteFile": {
            "type": "object",
            "required": [
                "file_id"
            ],
            "properties": {
                "file_id": {
                    "type": "integer"
                }
            }
        },
...
        "request.ParamGetFileDetailsByID": {
            "type": "object",
            "required": [
                "file_id"
            ],
            "properties": {
                "file_id": {
                    "type": "integer"
                }
            }
        },
...
        "request.ParamGetRelationFile": {
            "type": "object",
            "required": [
                "relation_id"
            ],
            "properties": {
                "relation_id": {
                    "type": "integer"
                }
            }
        },
...
        "request.ParamSendEmail": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
...
        "request.ParamUploadAccountAvatar": {
            "type": "object"
        }
    }
}